[rs-catalog] [Search] Handle facets. WIP.

+++ rs-catalog/search/search-rest/src/test/java/fr/cnes/regards/modules/search/rest/CatalogControllerIT.java
@@ -6,7 +6,7 @@ package fr.cnes.regards.modules.search.rest;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.After;
+import org.hamcrest.Matchers;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
@@ -65,22 +65,27 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
     /**
      * A dummy collection
      */
-    public static final Collection COLLECTION = new Collection(null, DEFAULT_TENANT, null);
+    public static final Collection COLLECTION = new Collection(null, DEFAULT_TENANT, "mycollection");
 
     /**
      * A dummy dataobject
      */
-    public static final DataObject DATAOBJECT = new DataObject(null, DEFAULT_TENANT, null);
+    public static final DataObject DATAOBJECT = new DataObject(null, DEFAULT_TENANT, "mydataobject");
 
     /**
      * A dummy dataset
      */
-    public static final Dataset DATASET = new Dataset(null, DEFAULT_TENANT, null);
+    public static final Dataset DATASET = new Dataset(null, DEFAULT_TENANT, "mydataset");
+
+    /**
+     * An other dummy dataset
+     */
+    public static final Dataset DATASET_1 = new Dataset(null, DEFAULT_TENANT, "mydataset");
 
     /**
      * A dummy document
      */
-    public static final Document DOCUMENT = new Document(null, DEFAULT_TENANT, null);
+    public static final Document DOCUMENT = new Document(null, DEFAULT_TENANT, "mydocument");
 
     /**
      * @throws java.lang.Exception
@@ -92,19 +97,16 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
 
         // Populate the ElasticSearch repository
         runtimeTenantResolver.forceTenant(DEFAULT_TENANT);
+        if (esRepository.indexExists(DEFAULT_TENANT)) {
+            esRepository.deleteIndex(DEFAULT_TENANT);
+        }
         esRepository.createIndex(DEFAULT_TENANT);
         esRepository.save(DEFAULT_TENANT, COLLECTION);
         esRepository.save(DEFAULT_TENANT, DATASET);
+        esRepository.save(DEFAULT_TENANT, DATASET_1);
         esRepository.save(DEFAULT_TENANT, DATAOBJECT);
         esRepository.save(DEFAULT_TENANT, DOCUMENT);
-    }
-
-    /**
-     * @throws java.lang.Exception
-     */
-    @After
-    public void tearDown() throws Exception {
-        esRepository.deleteIndex(DEFAULT_TENANT);
+        esRepository.refresh(DEFAULT_TENANT);
     }
 
     /**
@@ -114,7 +116,7 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
     @Purpose("Le système doit permettre de réaliser une recherche par critères sur l’ensemble du catalogue.")
     @Requirement("REGARDS_DSL_DAM_CAT_510")
     public final void testSearchAll() {
-        final List<ResultMatcher> expectations = new ArrayList<>(1);
+        final List<ResultMatcher> expectations = new ArrayList<>();
         expectations.add(MockMvcResultMatchers.status().isOk());
         RequestParamBuilder builder = RequestParamBuilder.build().param("q", CatalogControllerTestUtils.Q);
         performDefaultGet("/search", expectations, "Error searching all entities", builder);
@@ -126,8 +128,8 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
     @Test
     @Purpose("Le système doit permettre de fournir des facettes pour toute recherche dans le catalogue.")
     @Requirement("REGARDS_DSL_DAM_CAT_610")
-    public final void testSearchAll_withFacettes() {
-        final List<ResultMatcher> expectations = new ArrayList<>(1);
+    public final void testSearchAll_withFacets() {
+        final List<ResultMatcher> expectations = new ArrayList<>();
         expectations.add(MockMvcResultMatchers.status().isOk());
         RequestParamBuilder builder = RequestParamBuilder.build().param("q", CatalogControllerTestUtils.Q)
                 .param("facets", CatalogControllerTestUtils.FACETS_AS_ARRAY);
@@ -141,7 +143,7 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
     @Purpose("Le système doit permettre de manière synchrone d’accéder aux informations d’une collection via son IP_ID.")
     @Requirement("REGARDS_DSL_DAM_COL_310")
     public final void testGetCollection() {
-        final List<ResultMatcher> expectations = new ArrayList<>(1);
+        final List<ResultMatcher> expectations = new ArrayList<>();
         expectations.add(MockMvcResultMatchers.status().isOk());
         performDefaultGet("/collections/{urn}", expectations, "Error retrieving a collection", COLLECTION.getIpId());
     }
@@ -153,10 +155,16 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
     @Test
     @Purpose("Le système doit permettre de manière synchrone de rechercher des collections à partir de critères basés sur des éléments du modèle de données.")
     @Requirement("REGARDS_DSL_DAM_COL_410")
-    public final void testSearchCollections() {
-        final List<ResultMatcher> expectations = new ArrayList<>(1);
+    public final void testSearchCollections_shouldFindOneWithoutFacets() {
+        final List<ResultMatcher> expectations = new ArrayList<>();
         expectations.add(MockMvcResultMatchers.status().isOk());
-        RequestParamBuilder builder = RequestParamBuilder.build().param("q", CatalogControllerTestUtils.Q);
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT).isNotEmpty());
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".content", Matchers.notNullValue()));
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".content.[0].content.label",
+                                                        Matchers.is("mycollection")));
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".facets").doesNotExist());
+        RequestParamBuilder builder = RequestParamBuilder.build()
+                .param("q", CatalogControllerTestUtils.Q_FINDS_ONE_COLLECTION);
         performDefaultGet("/collections/search", expectations, "Error searching collections", builder);
     }
 
@@ -165,7 +173,7 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
      */
     @Test
     public final void testGetDataset() {
-        final List<ResultMatcher> expectations = new ArrayList<>(1);
+        final List<ResultMatcher> expectations = new ArrayList<>();
         expectations.add(MockMvcResultMatchers.status().isOk());
         performDefaultGet("/datasets/{urn}", expectations, "Error retrieving a dataset", DATASET.getIpId());
     }
@@ -175,10 +183,18 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
      * {@link fr.cnes.regards.modules.search.rest.CatalogController#searchDatasets(java.lang.String, org.springframework.data.domain.Pageable, org.springframework.data.web.PagedResourcesAssembler)}.
      */
     @Test
-    public final void testSearchDatasets() {
-        final List<ResultMatcher> expectations = new ArrayList<>(1);
+    public final void testSearchDatasets_shouldFindTwoWithoutFacets() {
+        final List<ResultMatcher> expectations = new ArrayList<>();
         expectations.add(MockMvcResultMatchers.status().isOk());
-        RequestParamBuilder builder = RequestParamBuilder.build().param("q", CatalogControllerTestUtils.Q);
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT).isNotEmpty());
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".content", Matchers.notNullValue()));
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".content.[0].content.label",
+                                                        Matchers.is("mydataset")));
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".content.[1].content.label",
+                                                        Matchers.is("mydataset")));
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".facets").doesNotExist());
+        RequestParamBuilder builder = RequestParamBuilder.build()
+                .param("q", CatalogControllerTestUtils.Q_FINDS_TWO_DATASETS);
         performDefaultGet("/datasets/search", expectations, "Error searching datasets", builder);
     }
 
@@ -189,8 +205,12 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
     @Purpose("Le système doit permettre de consulter les métadonnées d’un objet de données du catalogue.")
     @Requirement("REGARDS_DSL_DAM_CAT_470")
     public final void testGetDataobject() {
-        final List<ResultMatcher> expectations = new ArrayList<>(1);
+        final List<ResultMatcher> expectations = new ArrayList<>();
         expectations.add(MockMvcResultMatchers.status().isOk());
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT).isNotEmpty());
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".content", Matchers.notNullValue()));
+        expectations
+                .add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".content.label", Matchers.is("mydataobject")));
         performDefaultGet("/dataobjects/{urn}", expectations, "Error retrieving a dataobject", DATAOBJECT.getIpId());
     }
 
@@ -199,10 +219,16 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
      * {@link fr.cnes.regards.modules.search.rest.CatalogController#searchDataobjects(java.lang.String, java.util.List, org.springframework.data.domain.Pageable, org.springframework.data.web.PagedResourcesAssembler)}.
      */
     @Test
-    public final void testSearchDataobjects() {
-        final List<ResultMatcher> expectations = new ArrayList<>(1);
+    public final void testSearchDataobjects_shouldFindOneWithFacets() {
+        final List<ResultMatcher> expectations = new ArrayList<>();
         expectations.add(MockMvcResultMatchers.status().isOk());
-        RequestParamBuilder builder = RequestParamBuilder.build().param("q", CatalogControllerTestUtils.Q);
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT).isNotEmpty());
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".content", Matchers.notNullValue()));
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".content.[0].content.label",
+                                                        Matchers.is("mydataobject")));
+        expectations.add(MockMvcResultMatchers.jsonPath(JSON_PATH_ROOT + ".facets", Matchers.notNullValue()));
+        RequestParamBuilder builder = RequestParamBuilder.build()
+                .param("q", CatalogControllerTestUtils.Q_FINDS_ONE_DATAOBJECT);
         performDefaultGet("/dataobjects/search", expectations, "Error searching dataobjects", builder);
     }
 
@@ -212,7 +238,7 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
      */
     @Test
     public final void testSearchDataobjectsReturnDatasets() {
-        final List<ResultMatcher> expectations = new ArrayList<>(1);
+        final List<ResultMatcher> expectations = new ArrayList<>();
         expectations.add(MockMvcResultMatchers.status().isOk());
         RequestParamBuilder builder = RequestParamBuilder.build().param("q", CatalogControllerTestUtils.Q);
         performDefaultGet("/dataobjects/datasets/search", expectations, "Error searching datasets via dataobjects",
@@ -226,7 +252,7 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
     @Purpose("Le système doit permettre de manière synchrone d’accéder aux informations d’un document via son IP_ID.")
     @Requirement("REGARDS_DSL_DAM_DOC_310")
     public final void testGetDocument() {
-        final List<ResultMatcher> expectations = new ArrayList<>(1);
+        final List<ResultMatcher> expectations = new ArrayList<>();
         expectations.add(MockMvcResultMatchers.status().isOk());
         performDefaultGet("/documents/{urn}", expectations, "Error retrieving a document", DOCUMENT.getIpId());
     }
@@ -239,7 +265,7 @@ public class CatalogControllerIT extends AbstractRegardsTransactionalIT {
     @Purpose("Le système doit permettre de manière synchrone de rechercher des documents à partir de critères basés sur des éléments du modèle de données.")
     @Requirement("REGARDS_DSL_DAM_DOC_510")
     public final void testSearchDocuments() {
-        final List<ResultMatcher> expectations = new ArrayList<>(1);
+        final List<ResultMatcher> expectations = new ArrayList<>();
         expectations.add(MockMvcResultMatchers.status().isOk());
         RequestParamBuilder builder = RequestParamBuilder.build().param("q", CatalogControllerTestUtils.Q);
         performDefaultGet("/documents/search", expectations, "Error searching documents", builder);
